objetosProjeto/mvc/requirements.txt
flask
flask-restful
mysql-connector-python
python-dotenv